Refactor: Split useCourseData hook into smaller files
This commit refactors the `useCourseData.ts` hook into smaller, more manageable files to improve code organization and maintainability. The functionality remains unchanged.
